improve api

+++ src/main/java/ru/usque/pelican/repository/PelicanUserRepository.java
@@ -5,7 +5,7 @@ import ru.usque.pelican.entities.PelicanUser;
 
 import java.util.List;
 
-public interface PelicanUserRepository extends CrudRepository<PelicanUser, Long>  {
+public interface PelicanUserRepository extends CrudRepository<PelicanUser, Integer>  {
     List<PelicanUser> findByEmail(String email);
     PelicanUser findById(Integer id);
     List<PelicanUser> findByLogin(String login);
